update route: created RouteService and RouteResource

+++ src/main/java/com/transport/easy/transporteasy/repositories/RotaRepository.java
@@ -3,5 +3,5 @@ package com.transport.easy.transporteasy.repositories;
 import com.transport.easy.transporteasy.entities.Rota;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RotaRepository extends JpaRepository<Rota, Integer> {
+public interface RotaRepository extends JpaRepository<Rota, Long> {
 }
